refactor toasts list to new toast reducer with queue property
