+++ src/reducers/application.js
@@ -3,10 +3,6 @@ import { buildActions } from 'utils';
 export const types = buildActions('application', [
   'INIT_APP',
   'LOGIN_USER',
-  'POP_TOAST',
-  'ADD_TOAST',
-  'REMOVE_TOAST',
-  'HIDE_TOAST',
   'REQUEST_TOKEN',
   'REQUEST_TOKEN_SUCCESS',
   'REQUEST_TOKEN_FAILURE',
@@ -31,26 +27,6 @@ const loginUser = (emailAddress, password) => ({
   type: types.LOGIN_USER,
   emailAddress,
   password
-});
-
-const popToast = toast => ({
-  type: types.POP_TOAST,
-  toast
-});
-
-const addToast = toast => ({
-  type: types.ADD_TOAST,
-  toast
-});
-
-const removeToast = id => ({
-  type: types.REMOVE_TOAST,
-  id
-});
-
-const hideToast = id => ({
-  type: types.HIDE_TOAST,
-  id
 });
 
 const requestToken = (emailAddress, password) => ({
@@ -123,10 +99,6 @@ const registerUserFailure = error => ({
 export const actions = {
   initApp,
   loginUser,
-  popToast,
-  addToast,
-  removeToast,
-  hideToast,
   requestToken,
   requestTokenSuccess,
   requestTokenFailure,
@@ -157,42 +129,11 @@ export const initialState = {
     registering: false,
     complete: false,
     error: null
-  },
-  toasts: []
+  }
 };
 
 export const reducer = (state = initialState, action = {}) => {
   switch (action.type) {
-    case types.ADD_TOAST:
-      return {
-        ...state,
-        toasts: [...state.toasts, action.toast]
-      };
-    case types.REMOVE_TOAST:
-      return {
-        ...state,
-        toasts: state.toasts.filter(toast => toast.id !== action.id)
-      };
-    case types.HIDE_TOAST: {
-      const originalToast = state.toasts.find(toast => toast.id === action.id);
-
-      if (!originalToast) {
-        return state;
-      }
-
-      const newToast = {
-        ...originalToast,
-        show: false
-      };
-      const filteredToasts = state.toasts.filter(
-        toast => toast.id !== action.id
-      );
-
-      return {
-        ...state,
-        toasts: [...filteredToasts, newToast]
-      };
-    }
     case types.LOGIN_USER:
       return {
         ...state,
