+++ src/sagas/index.js
@@ -1,7 +1,8 @@
-import { fork } from 'redux-saga/effects';
+import { all, fork } from 'redux-saga/effects';
 
 import application from './application';
+import toast from './toast';
 
 export default function* saga() {
-  yield fork(application);
+  yield all([application, toast].map(fork));
 }
