+++ src/selectors/toast.js
@@ -1,0 +1,8 @@
+import { createSelector } from 'reselect';
+
+export const getToasts = state => state.toast;
+
+export const getQueue = createSelector(
+  getToasts,
+  toast => toast.queue
+);
